Add Alt as modifier key to git log search key binding (#17)
After much feedback around git log search's Ctrl+l key binding conflicting with clear screen, I have decided to change its key binding to Ctrl+Alt+l. An alternative option I considered was just Alt+l, which is shorter but I realized conflicts with __fish_list_current_token, which is actually quite useful and I plan on using it now that I've discovered it.

One huge problem with this change is that it is backwards incompatible and fisher has not provided me with a way to announce to users of fzf.fish that the key binding has changed. I had considered using a hack of checking if an arbitrary file existed, printing an error message if it doesn't and then creating that file, but it adds a lot of difficult-to-maintain code and slows down the shell startup.
Hopefully, people who discover that Ctrl+l no longer works will come to this repo and see this commit. I will also create a new release announcing this change. I am sorry for any confusion caused by this.