Fix file preview not working non non-Mac OSes (#21)
# Problem
File preview wasn't working on Linux and presumably Windows because the file command that ships with Mac, used to determine the file type, is outdated and has different options.

# Solution
Use ls to get the file type instead. Big thanks to @chshbh for the idea of using ls! He opened up a great PR that this is based on.