Convert license to markdown (#35)
Convert license to markdown. As with README, CONTRIBUTING, CHANGELOG and the rest of the family, LICENSE is meant to be read by humans (granted no one ever actually reads the MIT license because we all know what it says, but still). Anything that is meant to be read by humans should be rendered by the browser into good looking HTML, not displayed as plain text.