Put function arguments first before description
Arguments are more important and thus is makes sense to put them first so the person reading the code can get to them sooner.
FWIW I took a survey of several repos and all but one of them (of the ones that used both -a and -d flags) put the description last.
Resolves #38