Always pre-populate fzf's search query with the current token (#73)
Previously, only command history search used the current token from the commandline to pre-populate fzf's search query. Now, all functions do that. This should be a welcome change for all users. Not only does this make searching in fzf faster, but this should reduce some hassle. The previous behavior of simply inserting the user's selection into the commandline meant that if there was a current token when the user triggered fzf.fish, that token would become one word with their first selection. By replacing the current token instead, users no longer have to go back and delete it to clean up their command.

One more thing to note is that a special case was made for searching for shell variables. If the current token begins with a $, the $ is not overwritten when replacing the current token with the selected variable.