Limit file search to directory under cursor if it has a trailing / (#75)
This new behavior allows the user to control the directory in which the search file feature initiates its search. If the current token is a directory and it has a trailing slash (e.g. functions/<CURSOR>), then we assume the user wants to search for files in that directory.
This can be useful when in a directory containing large subdirectories (e.g. on macOS, the home directory contains ~/Library which is very large). Moreover, it also allows searching a parent or sibling directory by using ../ (e.g. ../Downloads/<CURSOR>. 
What's great is that this new behavior is that it complements tab completion. When a user wants to access a file in a subdirectory, they may start typing the name of the directory and hit TAB. When tab completion writes out the directory's path, it will already append a trailing slash for the user, allowing the user to quickly execute the search files feature and search within that directory.