Fix multi-line command bug
Before, selecting a multi-line command when using the search command history feature would result in the command lined being replaced with only the top line of the selected command.
For example, selecting this command from history...
  function example
      echo "This is just an example"
  end
...would result in the command line becoming...
  function example
with the rest of the command lost.

The bug occurred because when a command substitution has a multi-line output and is assigned to a variable, each line of the output is give its own index in the variable. So, referencing index 2 only gave us one line of the command. To fix this, we have to string collect the multi-line output into a single string before assigning it.
This fixes the bug by doing just that so users can now select the entirety of multi-line commands from history.