Move Prior Art section to Wiki
Similar to the Troubleshooting section, I've been having to udpdate the Prior Art section rather frequently and it would be much lighterweight if it were a Wiki page. Furthermore, I have been hindered from writing a more in-depth explanation of how fzf.fish's design philosophy is an improvement over jethrokuan/fzf's because I didn't want the readme to be hard to scroll through. Moving it to the readme will resolve that tension.