Allow passing custom fzf opts to individual commands (#140)
The following variables can store the custom options that will be passed to fzf by their respective feature:
  fzf_dir_opts => search directory
  fzf_git_status_opts => search git status
  fzf_git_log_opts => search git log
  fzf_history_opts => search history
  fzf_shell_vars_opts => search shell variables

They are always appended last to fzf's argument list so that they can override any hardcoded fzf options. This might lead to some users shooting themselves in the foot but also unlocks numerous opportunities for the user to customize or augment the existing features.

This was an oft requested feature that I parried for a long time but now I am very excited to finally add. With this, change, some of the cool ideas now made possible include:
- add key bindings within fzf to operate on the selected line, e.g. open file in vim, preview file as image, copy to clipboard, git checkout commit, git reset commit
- adjust the preview window or command
- re-populate fzf's input list on demand
- change the search mode
- and many, many more