[Search variables] enable selecting multiple variables (#165)
Enable and support selecting multiple variables by passing --multi to fzf and string joining the results. This will be useful, for example, when needing to select multiple related variables to erase, show, query. Search variables continues to preserve the $ if it's in the current token and prepends it each selected variable. This might be rarely used but will be useful if selecting multiple related variables to echo or write to a file.