Give search directory and search variables a simpler, canonical function name
I've wanted to rename these functions for a while now ever since I
decided to refer to them in the documentation as [Search directory] and
[Search variables]. These names are not only shorter but now directly
map to what they do. Futhermore, all the other function names followed
the same standard. However, I could not make this change until I
overhauled custom key bindings, as it would have broken custom key
bindings for users and would require some manual and tedious effort to
migrate. Now that users have to migrate to custom key bindings anyway,
and it is easy to do, this is the opportune time to make this change.