[Search directory] Hide ./ prefix when fd version >= 8.3.0 (#211)
fd >= 8.3.0 prefixes ./ to all paths when the output is being piped rather than sent to TTY. This is ugly and we don't need this so we disable this new behavior by passing a new flag --strip-cwd-prefix if the version is >= 8.3.0. Eventually, we will require fd 8.3.0 as a minimum version and always pass that flag.