Introduce feature to search for processes (#216)
Useful for commands such as kill -STOP $pid or lsof -p $pid

Includes a preview that shows other useful information about the process selected such as parent pid, CPU use, memory use, and start time.

There will be a follow up commit to this one to clean up the code, add tests, add a gif, and enable fzf multi-select. I'm merging this in early because the PR has been open for far too long and is already in a usable, tested state.