[Search directory] Prevent duplicate / when fd version >= 8.4.0 (#241)
fd 8.4.0 contains this change: "Directories are now printed with an additional path separator at the end".
fzf.fish has another feature that appends / for quick cds. This causes a duplicate / at the end that is confusing and looks terrible, though it may not actually cause issues. We fix this by preventing the / to be appended when the fd version is >= 8.4.0.