[Search git status] rewrite parsing logic without regex and fix bug with renames
There was a bug in the status parsing logic in which the unstaged output header would be printed if a file was renamed in the index but had no unstaged changes. I decided the regex code was too hard to read and unwieldy--hence how the bug survived code review--so rewrote it using simple, self-documenting substring checks while fixing the bug. This should also marginally improve performance.