[Search git status] Preview only changes since file renamed (#263)
Previously, the git diff preview for a renamed path shows the entire file as being added. A more useful diff is to view only the modifications made to the file content. To accomplish this, we need a special condition for renames to diff the current (post-rename) path with the original path.

In addition, this PR fixes git status previews for renamed paths, which was accidentally broken in 38896f2.